MVC pattern implemented
View does not refer to $scope for each individual model object instance.
Model is now a factory managed object, a complex structure describing
the state of the application. Its lifecycle is managed within app.js.
Views refer to $scope.model, simplifying a lot of code in controllers. I
think this approach is way cleaner, as opposed to misunderstanding
$scope itself being a model which is not.