Project reorganized
I prefer to have a "src" folder for application sources, and "dist"
folder for the build process outputs. This setup is more inline with
typical compiled langs projects like java. Although JS is a scripting
language - tools and frameworks around it make it feasible to arrange
projects like this. Been seeing this a standard too in high quality
projects like Angular or Angular Material, so applying what the best are
doing already.